JSON multihash is string, not object, in C++ #153
