+++ programs/koinos-types/lang/koinos_codegen_cpp/rt/json.hpp
@@ -2,6 +2,7 @@
 #include <koinos/pack/rt/json_fwd.hpp>
 
 #include <koinos/pack/rt/exceptions.hpp>
+#include <koinos/pack/rt/multihash_json.hpp>
 #include <koinos/pack/rt/opaque.hpp>
 #include <koinos/pack/rt/reflect.hpp>
 #include <koinos/pack/rt/typename.hpp>
@@ -410,26 +411,6 @@ inline void from_json( const json& j, optional< T >& v, uint32_t depth )
       from_json( j, tmp, depth );
       v = std::move( tmp );
    }
-}
-
-// multihash
-inline void to_json( json& j, const multihash& v )
-{
-   json tmp;
-   to_json( tmp, v.digest );
-   j[ "hash" ] = v.id;
-   j[ "digest" ] = std::move( tmp );
-}
-
-inline void from_json( const json& j, multihash& v, uint32_t depth )
-{
-   if( !(j.is_object()) ) throw json_type_mismatch( "Unexpected JSON type: object exptected" );
-   if( !(j.size() == 2) ) throw json_type_mismatch( "Multihash JSON type must only contain two fields" );
-   if( !(j.contains( "hash" )) ) throw json_type_mismatch( "Multihash JSON type must contain field 'hash'" );
-   if( !(j.contains( "digest" )) ) throw json_type_mismatch( "Multihash JSON type must contain field 'digest'" );
-
-   v.id = j[ "hash" ].get< uint64_t >();
-   from_json( j[ "digest" ], v.digest );
 }
 
 template< typename T >
